requirements.txt
fastapi
gensim
httpx
uvicorn 
pandas 
numpy 
openai==0.28.0
plotly
pydantic 
pymongo
pytest
pytest-asyncio
python-dotenv
ipykernel
matplotlib
nbformat
nltk
seaborn
scipy==1.10.1
transformers 
torch 
scikit-learn
wordcloud